pstring abstract
